data_service/requirements.txt
Django==5.2.4
djangorestframework==3.15.1
